Move utility to the localization
